// types/forms.ts
export interface ContactFormData {
  fullName: string;
  email: string;
  phone: string;
  subject: string;
  message: string;
}

export interface EmployerEnquiryFormData {
  fullName: string;
  company: string;
  country: string;
  workEmail: string;
  phone: string;
  industry: string;
  requiredWorkforce: string;
  approximateWorkersCount: string;
  message: string;
}